+++ src/Phobos.h
@@ -89,6 +89,7 @@ public:
 		static bool ShowPowerDelta;
 		static bool ShowHarvesterCounter;
 		static bool ShowWeedsCounter;
+		static bool HideLightFlashEffects;
 	};
 
 	class Misc